requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-CORS==3.0.10
SQLAlchemy==1.4.46
psycopg2-binary==2.9.9
python-dotenv==0.21.1
Flask-Admin==1.6.1
gunicorn==21.2.0
marshmallow==3.20.1
Flask-Migrate==4.0.5